midi router handles ump midi messages

+++ include/interpolation/interpolation.h
@@ -5,7 +5,6 @@
 #include <sys/types.h>
 
 namespace interpolation {
-
-double interpolate(const parameters::parameter &parameter, u_int8_t control_value);
-
+double interpolate(const parameters::parameter &parameter,
+                   uint32_t control_value);
 } // namespace interpolation
